refactor fractal
