+++ src/fractal/fractal_types/mod.rs
@@ -1,0 +1,3 @@
+pub mod iterated_sin_z;
+pub mod julia_descriptor;
+pub mod mandelbrot;
