+++ src/fractal/mod.rs
@@ -1,0 +1,2 @@
+pub mod fractal;
+pub mod fractal_types;
