--- src/messages/fractal/fractal_types/mod.rs
@@ -1,1 +1,0 @@
-pub mod fractal_types;
