--- src/messages/fractal/mod.rs
@@ -1,1 +1,0 @@
-pub mod fractal_types;
