+++ src/messages/mod.rs
@@ -2,4 +2,3 @@ pub mod fragment_request;
 pub mod fragment_task;
 pub mod fragment_result;
 pub mod complementary_types;
-pub mod fractal;
